% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{excluded_tracks}
\alias{excluded_tracks}
\title{Excluded storm tracks for Atlantic basin storms}
\format{A data frame with 9,124 rows and 6 variables:
\describe{
  \item{storm_id}{Unique storm identifier with the storm name and year,
                 separated by a hyphen (e.g., "Alberto-1988",
                 "Katrina-2005"). Some storms are "Unnamed" or have other
                 generic names. In these cases, it may be preferable to identify
                 the storm using the first four characters of their
                 ATCF ID (\code{usa_atcf_id}) and the storm season rather than
                 the storm ID (e.g., 'AL11-2018').}
  \item{usa_atcf_id}{Character string with the Automated Tropical Cyclone
                 Forecasting System ID for the storm.}
  \item{date}{Character string with date and time of storm track recording, in
              the Universal Time Coordinated (UTC) time zone. This date is formated
              as \code{\%Y\%m\%d\%H\%M}.}
  \item{latitude}{Latitude of storm center, in decimal degrees}
  \item{longitude}{Longitude of storm center, in decimal degrees (note:
                   longitudes are given as negative values for
                   longitudes west of the prime meridian)}
  \item{wind}{1-minute maximum sustained surface wind speed, measured at
             10 meters above the ground, in knots (values are rounded to
             the nearest 5-knot value)}
}}
\source{
\url{https://www.nhc.noaa.gov/data/hurdat/hurdat2-1851-2018-120319.txt}
}
\usage{
excluded_tracks
}
\description{
A dataset containing the storm tracks for Atlantic basin tropical
storms between 1988 and 2018, from the US National Hurricane Center's
Best Track Atlantic hurricane database (HURDAT2). Only storms that
were excluded from the main 'hurr_tracks' dataset, because they did not
come within 250 km of at least one US county, are included.
}
\references{
Landsea, CW and JL Franklin, 2013. Atlantic Hurricane Database Uncertainty and
Presentation of a New Database Format. Monthly Weather Revue, 141, 3576--3592.
}
\keyword{datasets}
