% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ext_tracks_wind}
\alias{ext_tracks_wind}
\title{Extended best tracks county wind speeds for historical storms}
\format{A dataframe with 162,928 rows and 5 variables:
\describe{
  \item{fips}{County's 5-digit Federal Information Processing Standard (FIPS)
             code}
  \item{vmax_gust}{Maximum modeled gust wind speed in the county during
             the storm.}
  \item{vmax_sust}{Maximum modeled sustained wind speed in the county during
             the storm.}
  \item{sust_dur}{Minutes sustained wind was above 34 knots in the county during
             the storm.}
  \item{storm_id}{Unique storm identifier with the storm name and year,
             separated by a hyphen (e.g., "Alberto-1988",
             "Katrina-2005").}
}}
\source{
HURDAT2 data were obtained from:
\url{https://www.nhc.noaa.gov/data/}
}
\usage{
ext_tracks_wind
}
\description{
A dataframe with storm winds based on extended best tracks hurricane
data for historical Atlantic basin storms.
}
\details{
To calculate the wind values in this dataset, we used the wind radii from
   the HURDAT2 dataset for Atlantic-basin tropical storms (see the
   reference and source). These wind radii provide the radius in each of four quadrants
   (northeast, southeast, southwest, and northwest) from the storm's center for each
   of the 6-hour storm observations. We interpolated this data to every 15 minutes
   using linear interpolation (Section 3.3. of Press et al. (1992))
   and scaled each radius to 0.85 of its full value (the original radius gives the
   maximum extent of 64-knot, 50-knot, and 34-knot winds in each quadrant). For each
   15-minute point along the storm's track, we determined which counties fell within
   each wind radius. We then determined, for each county, the maximum wind value
   over the course of the storm. These wind values are limited to 34 knots, 50 knots,
   and 64 knots; if a county is reported as having 64-knot winds in this dataset, it
   therefore means that the county fell within 0.85 times the 64-knot maximum wind
   radius at least once during the storm, and so can be interpreted as the county
   likely experiencing at least 64-knot winds during the course of the storm.

   Gust wind speed was calculated using a gust factor of 1.49 applied to the estimated
   sustained wind speed (see the "Details" vignette of the \code{stormwindmodel}
   package for more details on the choice of this gust factor). Durations were based
   on the number of minutes winds were above 34 knots in the county over the course of
   the storm.
}
\references{
Landsea CW, Franklin JL, Beven JL. May 2015. The revised Atlantic
hurricane database (HURDAT2). United States National Oceanic and Atmospheric Administration's
National Weather Service.
\url{https://www.nhc.noaa.gov/data/hurdat/hurdat2-format-atlantic.pdf}
Retrieved November 28, 2019.

Press WH, Teukolsky SA, Vetterling WT, Flannery BP. 1992. Numerical Recipes in Fortran 77:
The Art of Scientific Computing. Cambridge University Press.
}
\author{
Andrea Schumacher \email{andrea.schumacher@colostate.edu}
}
\keyword{datasets}
