% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hurr_tracks}
\alias{hurr_tracks}
\title{Storm tracks for Atlantic basin storms}
\format{A data frame with 4,351 rows and 5 variables:
\describe{
  \item{storm_id}{Unique storm identifier with the storm name and year,
                 separated by a hyphen(e.g., "Alberto-1988",
                 "Katrina-2005")}
  \item{date}{Character string with date and time of storm track recording, in
              the Universal Time Coordinated (UTC) time zone. This date is formated
              as \code{\%Y\%m\%d\%H\%M}.}
  \item{latitude}{Latitude of storm center, in decimal degrees}
  \item{longitude}{Longitude of storm center, in decimal degrees (note:
                   longitudes are given as negative values for
                   longitudes west of the prime meridian)}
  \item{wind}{1-minute maximum sustained surface wind speed, measured at
             10 meters above the ground, in knots (values are rounded to
             the nearest 5-knot value)}
}}
\source{
\url{http://rammb.cira.colostate.edu/research/tropical_cyclones/tc_extended_best_track_dataset}
}
\usage{
hurr_tracks
}
\description{
A dataset containing the storm tracks for Atlantic basin tropical
storms between 1988 and 2015, from the Extended Best Track Dataset for
the Atlantic basin. Only storms that came within 250 km of at least
one US county are included.
}
\references{
Demuth J, DeMaria M, Knaff JA, 2006. Improvement of advanced microwave
sounder unit tropical cyclone intensity and size estimation algorithms.
Journal of Applied Meteorology and Climatology 45:1573-1581
}
\keyword{datasets}
