% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{excluded_tracks}
\alias{excluded_tracks}
\title{Excluded storm tracks for Atlantic basin storms}
\format{A data frame with 17,918 rows and 5 variables:
\describe{
  \item{storm_id}{Unique storm identifier with the storm name and year,
                 separated by a hyphen(e.g., "Alberto-1988",
                 "Katrina-2005"). For unnamed storms, the identifier uses the
                 first four digits of the storm's Automated Tropical Cyclone
                 Forecasting System ID and year, separated by a hyphen (e.g.,
                 "AL10-2007" for Tropical Depression Ten (unnamed storm) in 2007.)}
  \item{date}{Character string with date and time of storm track recording, in
              the Universal Time Coordinated (UTC) time zone. This date is formated
              as \code{\%Y\%m\%d\%H\%M}.}
  \item{latitude}{Latitude of storm center, in decimal degrees}
  \item{longitude}{Longitude of storm center, in decimal degrees (note:
                   longitudes are given as negative values for
                   longitudes west of the prime meridian)}
  \item{wind}{1-minute maximum sustained surface wind speed, measured at
             10 meters above the ground, in knots (values are rounded to
             the nearest 5-knot value)}
}}
\source{
\url{https://www.ncdc.noaa.gov/ibtracs/index.php?name=ib-v4-access}
}
\usage{
excluded_tracks
}
\description{
A dataset containing the storm tracks for Atlantic basin tropical
storms between 1988 and 2018, from the International Best Track Archive
for Climate Stewardship (IBTrACS) for the Atlantic basin. Only storms that
were excluded from the main 'hurr_tracks' dataset, because they did not
come within 250 km of at least one US county, are included.
}
\references{
Knapp KR, Kruk MC, Levinson DH, Diamond HJ, and Neumann CJ, 2010. The
International Best Track Archive for Climate Stewardship (IBTrACS).
Bulletin of the American Meteorological Society 91 (3), 363--376.
}
\keyword{datasets}
