% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ext_tracks_wind}
\alias{ext_tracks_wind}
\title{Extended best tracks county wind speeds for historical storms}
\format{A dataframe with 162,928 rows and 6 variables:
\describe{
  \item{fips}{County's 5-digit Federal Information Processing Standard (FIPS)
             code}
  \item{vmax_gust}{Maximum modeled gust wind speed in the county during
             the storm.}
  \item{vmax_sust}{Maximum modeled sustained wind speed in the county during
             the storm.}
  \item{gust_dur}{Time gust wind was above 20 m / s in the county during
             the storm.}
  \item{sust_dur}{Time sustained wind was above 20 m / s in the county during
             the storm.}
  \item{storm_id}{Unique storm identifier with the storm name and year,
             separated by a hyphen(e.g., "Alberto-1988",
             "Katrina-2005")}
}}
\usage{
ext_tracks_wind
}
\description{
A dataframe with storm winds based on extended best tracks hurricane
data for historical Atlantic basin storms.
}
\keyword{datasets}

